Fix roleLavel to UInt16

+++ Sources/TraderUserDto/Response/AdminDto.swift
@@ -8,7 +8,7 @@
 import struct Foundation.Date
 import struct Foundation.UUID
 
-/// User public data
+/// User public data.
 public struct AdminDto: Codable, Sendable {
 
     // MARK: Stored properties
@@ -19,7 +19,7 @@ public struct AdminDto: Codable, Sendable {
     public var firstName: String
     public var lastName: String?
 	public var lang: Lang
-    public var roleLevel: UInt8
+    public var roleLevel: UInt16
 	public var siwaIdentifier: String?
     public var tariff: Tariff
 	public var tryAt: Date
@@ -37,7 +37,7 @@ public struct AdminDto: Codable, Sendable {
         firstName: String = "",
         lastName: String? = nil,
 		lang: Lang = .en,
-		roleLevel: UInt8 = 1,	// client
+		roleLevel: UInt16 = 1,	// client
 		siwaIdentifier: String? = nil,
 		tariff: Tariff = .free,
 		tryAt: Date = .init(),
